requirements.txt
canonicalwebteam.flask-base==0.6.3
canonicalwebteam.discourse==2.0.0
